--- core/logging/config/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    id(ThunderbirdPlugins.Library.kmp)
-}
-
-android {
-    namespace = "net.thunderbird.core.logging.config"
-}
-
-kotlin {
-    sourceSets {
-        commonMain.dependencies {
-            api(projects.core.logging.api)
-            implementation(projects.core.logging.implComposite)
-            implementation(projects.core.logging.implFile)
-        }
-    }
-}
