+++ core/logging/impl-legacy/src/androidMain/kotlin/net/thunderbird/core/logging/legacy/DebugLogConfigurator.kt
@@ -1,17 +1,19 @@
-package net.thunderbird.core.logging.config
+package net.thunderbird.core.logging.legacy
 
 import net.thunderbird.core.logging.composite.CompositeLogSink
 import net.thunderbird.core.logging.file.FileLogSink
+import timber.log.Timber
+import timber.log.Timber.DebugTree
 
+// TODO: Implementation https://github.com/thunderbird/thunderbird-android/issues/9573
 class DebugLogConfigurator(
     private val syncDebugCompositeSink: CompositeLogSink,
     private val syncDebugFileLogSink: FileLogSink,
 ) {
     fun updateLoggingStatus(isDebugLoggingEnabled: Boolean) {
-        syncDebugCompositeSink.manager.removeAll()
+        Timber.uprootAll()
         if (isDebugLoggingEnabled) {
-            syncDebugCompositeSink.manager.add(syncDebugCompositeSink)
-            syncDebugCompositeSink.manager.add(syncDebugFileLogSink)
+            Timber.plant(DebugTree())
         }
     }
