+++ legacy/core/src/main/java/com/fsck/k9/preferences/KoinModule.kt
@@ -3,7 +3,7 @@ package com.fsck.k9.preferences
 import com.fsck.k9.Preferences
 import kotlin.time.ExperimentalTime
 import net.thunderbird.core.android.account.LegacyAccountDtoManager
-import net.thunderbird.core.logging.config.DebugLogConfigurator
+import net.thunderbird.core.logging.legacy.DebugLogConfigurator
 import net.thunderbird.core.preference.DefaultPreferenceChangeBroker
 import net.thunderbird.core.preference.GeneralSettingsManager
 import net.thunderbird.core.preference.PreferenceChangeBroker
